# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000/v1"
    description: "Server locale (v1)"
info:
  title: WASAText API
  description: Struttura API del progetto WASAText del corso di Programmazione Web.
  version: 1.0.0
...
security:
   - LegacyAuth: []
paths:
  /login:
    post:
      tags: ["login"]
      summary: login dell'utente
      description: accesso dell'utente tramite username; se l'utente non esiste viene creato
      operationId: doLogin
      security: []
      requestBody:
        description: dettagli utente
        required: true
        content:
          application/json:
            schema:
              description: nome utente richiesto per autenticazione/registrazione
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: Giorgia
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  description: nome utente
      responses:
        '200':
          description: utente riconosciuto e loggato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    identifier: 125461561
        '201':
          description: utente creato e loggato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                created:
                  value:
                    identifier: 125461561
...
  /groups:
    post:
      tags: ["groups"]
      summary: crea un nuovo gruppo con membri iniziali
      description: crea un nuovo gruppo con membri iniziali e creatore compresi, selezionando titolo e foto di gruppo
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupname, membersList] 
              description: oggetto che ritorna le componenti di costruzione di un gruppo
              properties:
                groupname: 
                  type: string
                  description: Nome del gruppo
                  minLength: 1
                  maxLength: 100
                  pattern: '^.*?$'
                groupPhoto: 
                  type: string
                  format: url
                  minLength: 10
                  maxLength: 200
                  pattern: '^.*?$'
                  description: Foto del gruppo (URL)
                membersList: 
                  type: array
                  minItems: 1
                  maxItems: 2000
                  items:
                    type: integer
                  description: lista di utenti (il creatore è incluso automaticamente)
      responses:
        '201':
          description: gruppo creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                success:
                  value:
                    Group:
                      id: 5161545165
                      groupname: Fantacalcio 2025
                      groupPhoto: "https://example.com/photo.jpg" 
                      membersList:  [23413453421, 53344234531, 453123453, 453123453]
  /groups/{groupId}/members:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Identificatore numerico del gruppo
        schema:
          type: integer
          minimum: 1
    get: 
      tags: ["groups"]
      summary: ottiene la lista dei membri del gruppo
      description: restituisce la lista completa degli utenti che fanno parte del gruppo
      operationId: getGroupMembers
      responses:
        '200':
          description: lista membri ottenuta con successo
          content:
            application/json:
              schema:
                type: object
                description: Oggetto contenente la lista dei membri del gruppo
                properties:
                  membersList:
                    type: array
                    description: Lista degli utenti che fanno parte del gruppo
                    minItems: 0
                    maxItems: 2000
                    items:
                      $ref: '#/components/schemas/User'
              examples:
                success:
                  value:
                    membersList:
                      - id: 101
                        username: "Mario"
                        profilePhoto: "https://example.com/mario.jpg"
                      - id: 102
                        username: "Luigi"
                        profilePhoto: "https://example.com/luigi.jpg"

    post:
      tags: ["groups"]
      summary: aggiunge membri al gruppo
      description: un membro aggiunge nuovi membri ad un gruppo esistente tramite username
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username] 
              description: oggetto che ritorna lo username dell'utente da aggiungere
              properties:
                username:
                  description: Username dell'utente da aggiungere 
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: '^.*?$'
      responses:
        '200':
          description: membri aggiunti con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                success:
                  value:
                    Group:
                      id: 5161545165
                      groupname: Fantacalcio 2025
                      groupPhoto: "https://example.com/photo.jpg" 
                      membersList:  [23413453421, 53344234531, 453123453, 453123453, 99999]
        '403':
          description: utente non appartenente al gruppo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: membro da aggiungere non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/leave:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Identificatore numerico del gruppo
        schema:
          type: integer
          minimum: 1
    delete:
      tags: ["groups"]
      summary: lascia un gruppo
      description: l'utente corrente abbandona un gruppo esistente di cui è membro
      operationId: leaveGroup
      responses:
        '204':
          description: abbandono effettuato con successo
  /groups/{groupId}/name:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Identificatore numerico del gruppo
        schema:
          type: integer
          minimum: 1
    put:
      tags: ["groups"]
      summary: aggiorna il nome del gruppo
      description: un membro aggiorna il nome di un gruppo esistente con uno nuovo
      operationId: setGroupName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupname] 
              description: oggetto che ritorna il nuovo nome da dare al gruppo
              properties:
                groupname: 
                  description: Il nuovo nome del gruppo
                  type: string
                  minLength: 1
                  maxLength: 100
                  pattern: '^.*?$'
      responses:
        '200':
          description: nome del gruppo aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                success:
                  value:
                    Group:
                      id: 5161545165
                      groupname: "Nuovo Nome Gruppo"
                      groupPhoto: "https://example.com/photo.jpg" 
                      membersList:  [23413453421, 53344234531]
        '403':
          description: utente non appartenente al gruppo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/photo:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Identificatore numerico del gruppo
        schema:
          type: integer
          minimum: 1
    put:
      tags: ["groups"]
      summary: aggiorna la foto del gruppo
      description: un membro aggiorna la foto di un gruppo esistente caricando un file
      operationId: setGroupPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupPhoto] 
              description: oggetto che ritorna la foto da caricare per il gruppo
              properties:
                groupPhoto: 
                  description: URL o dati della nuova foto
                  type: string
                  format: url
                  minLength: 10
                  maxLength: 200
      responses:
        '200':
          description: foto del gruppo aggiornata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                success:
                  value:
                    Group:
                      id: 5161545165
                      groupname: Fantacalcio 2025
                      groupPhoto: "https://example.com/nuova-foto.jpg"
                      membersList:  [23413453421, 53344234531]
        '403':
          description: utente non appartenente al gruppo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
...
  schemas:
    LoginResponse:
      type: object
      description: Risposta contenente l'identificatore dell'utente
      properties:
        identifier:
          type: integer
          description: Identificatore univoco dell'utente
          example: 125461561
    User:
      type: object
      description: Rappresenta un utente del sistema
      properties:
        id:
          type: integer
          description: Identificatore utente
        username:
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: Nome utente unico
        profilePhoto:
          type: string
          format: url
          description: URL verso la foto profilo
        token:
          type: string
          description: Token di sessione per l'autenticazione
    Group:
      type: object
      description: Rappresenta un gruppo di chat
      properties:
        id:
          type: integer
          description: Identificatore del gruppo
        groupname:
          type: string
          description: Nome visualizzato del gruppo
          pattern: "^.*$"
        groupPhoto:
          type: string
          format: url
          description: URL della foto del gruppo
        membersList:
          type: array
          minItems: 1
          maxItems: 2000
          items:
            type: integer
          description: Lista di id utenti membri del gruppo
...
    Error:
      type: object
      description: schema di errore generico
      properties:
        code:
          type: integer
          description: Codice di errore
        message:
          type: string
          minLength: 5
          maxLength: 200
          pattern: "^.*$"
          description: messaggio di spiegazione dell'errore
  securitySchemes: 

    LegacyAuth:
      type: apiKey
      name: sessionToken
      in: header
      description: valore del token inviato dal server nel momento di login